backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2

python-multipart
python-dotenv
pydantic-settings
google-generativeai
